+++ src/GameLibrary/Graphics/Color.cpp
@@ -121,10 +121,10 @@ namespace GameLibrary
 
 	byte Color_compositeByte(byte orig, byte comp)
 	{
-		/*float x = (float)orig;
-		float n = (float)comp;
+		/*double x = (double)orig;
+		double n = (double)comp;
 		return (byte)((-n / 255) * (n - x - 255));*/
-		return (byte)(orig*((float)comp/255));
+		return (byte)(orig*((double)comp/255));
 	}
 
 	Color Color::composite(const Color&comp) const
